chore: restructure write operation into module
